Literature Review
Submission of a close to finalised draft of the literature review which has already been sent to mentor to be reviewed. Past tense and passive writing style have been adhered to with 2/3 of the recommended word count being met. This will allow room for further information to be added as required. In line citations for all information presented have been done with correct referencing in Harvard style compiled at the end of the document. Figures were added were deemed necessary to help clarify certain technicalities. Literature includes an overview of IoT, WSNs and LPWANS. After which literature focuses on two key variations of LPWANs listing two proponent technologies for each variation. Identifying that LoRa may be the most suitable technology for the study LoRa and LoRaWAN are expanded on.