Results and Discussion
Commit includes documentation for results and discussion. Result relating to two main sets of tests were provided and discussed. Test for LoRaWAN security and testing of interfacing application layer protocols such as MQTT and HTTP. A table for comparison was also presented for the second set of tests. Recommended word limit has not been entirely met as additional discussion needs to be noted. References and correlations to literature have been made in context. Past tense and no first-person references recommendations were adhered to for the writing style of this section.