Conclusion, Introduction and Abstract
Commit includes the remaining sections required for the dissertation. Passive writing style will be adhered to on this document with only further work and recommendations being written in present or continuous tenses. Introduction has not been complied up until this point but has been clearly conceptualised and will be containing a definition of the problem, related scenario, and a clear description of personal motivation for the research. All research objectives will be defined along with the associated chapters in the study. Conclusion has been started but has not been finalised and will contain a summery and evaluation of the study’s findings along with recommendations and future work. These recommendations will be based around the need to research technologies that will be capable of better scalability. Abstract has not been compiled up until this commit but will be a concise summery of the study with insightful highlights.